design: Make thead content fix at the top.
The thead content in 10 tables are not fixed at the top, that results
in moving them with the tbody content. To fix this, this commit changes
the position property of thead to sticky at top.
Also this commit changes the color scheme of 10 tables that makes them
similar to the recent_topic_table.